@@ -1,11 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { HashRouter } from 'react-router-dom';
+import ReactDOM from 'react-dom/client';
 import App from './components/App';
+import { BrowserRouter } from 'react-router-dom';
 
-ReactDOM.render(
-  <HashRouter>
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <BrowserRouter>
     <App />
-  </HashRouter>,
-  document.getElementById('root')
+  </BrowserRouter>
+
 );
